Add files via upload
Regression test. Looks into tests/invoices directory, extracts the file list of utility bill pdfs and sends each file to the Utility Bill Processor for processing and data extraction. Extracted data is compared to the corresponding filename in the 'expected' directory, and all matches/mismatches are reported in the console. A total list of mismatches is reported at the end.